requirements.txt
google.generativeai
streamlit
dotenv
pillow
datetime